requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
boto3==1.26.156
botocore==1.29.156
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
distro==1.8.0
frozenlist==1.3.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
jmespath==1.0.1
multidict==6.0.4
openai==0.28.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
tqdm==4.65.0
typing_extensions==4.8.0
urllib3==1.26.16
yarl==1.9.2
